pyproject.toml
[project]
name = "multiplium"
version = "0.1.0"
description = "Deep research orchestration platform leveraging multi-provider LLM agent SDKs."
authors = [{ name = "Multiplium Engineering" }]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.7",
    "typer>=0.12",
    "rich>=13.7",
    "structlog>=24.1",
    "pandas>=2.2",
    "anthropic>=0.29",         # Validate version for Agent SDK features
    "openai-agents>=0.2",      # Confirm package name from GitHub release
    "google-genai>=0.3",       # Confirm version supporting tool execution
    "httpx>=0.27",
    "aiohttp>=3.9",
    "tenacity>=8.4",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-pyyaml",
    "types-requests",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

